VirtualShop/VirtualShop/src/Application/Products/Commands/UpdateProduct/UpdateProduct.cs
using VirtualShop.Application.Common.Models;

namespace VirtualShop.Application.Products.Commands.UpdateProduct;

public class UpdateProductCommandHandler : IRequestHandler<UpdateProductCommand, Result>
{
    private readonly IProductRepository _productRepository;
    private readonly IMapper _mapper;

    public UpdateProductCommandHandler(IProductRepository productRepository, IMapper mapper)
    {
        _productRepository = productRepository;
        _mapper = mapper;
    }

    public async Task<Result> Handle(UpdateProductCommand request, CancellationToken cancellationToken)
    {
        var product = await _productRepository.GetByIdAsync(request.Id);
        if (product is null)
        {
            return Result.Failure(["no such product!"]);
        }
        _mapper.Map( request, product);
        await _productRepository.UpdateAsync(product);
        return Result.Success();
    }
}
